requirements.txt
flask
werkzeug
nltk
scikit-learn
numpy
scipy
pdfplumber
python-pptx